// src/config/interfaces/interfaces.ts
export interface IAuthor {
  id: number;
  author_name: string;
  author_status: boolean;
}

export type TStory =
  | "Video Story"
  | "Story"
  | "video_story"
  | "web_story"
  | "add_story";
// ...
export interface TDisplayTag {
  name: string;
  slug: string;
  id: number;
  tagdescription: string;
  meta_title: string;
  meta_descriptions: string;
}
// ...
export interface ICategory {
  id: number;
  name: string;
  slug: string;
  meta_title: string;
  watermark: string;
  meta_discriptions: string;
}
// ...
export interface ITag {
  name: string;
  slug: string;
  id: number;
  tagtitle: string;
  tagdescription: string;
}

export interface IStoryDetails {
  title: string;
  alternative_title: string;
  watermark: string;
  excerpt: string;
  sub_heading: string;
  highlight: string;
  content?: string;
  content2?: string;
  blog_image?: {
    download_url: string;
    caption: string;
  };
  alternative_blog_image?: {
    download_url: string;
    caption: string;
  };
  photo_thumbnail?: {
    download_url: string;
    caption: string;
  };
  video_thumbnail?: {
    url: string;
  };
  blog_authors: IAuthor[];
  tags: ITag[];
  categories: ICategory[];
  subcategories: ICategory[];
  meta: {
    first_published_at: string;
    last_published_at: string;
    type: "blog.PhotoStory";
    search_description: string;
    seo_title: string;
  };
  story_type: TStory;
  news_slug: string;
  display_tags?: TDisplayTag[];
  is_live_news?: boolean;

  gallery_images?: {
    image: {
      meta: {
        download_url: string;
      };
      title: string;
    };
    caption: string;
    content: string;
  }[];

  live_news_content_story?: {
    id: number;
    title: string;
    image?: {
      url: string;
      width: number;
      height: number;
    };
    content: string;
    image_captions: string;
    created_at: string;
  }[];
}